refactor(cards.ts): finished refactoring store to pinia and tested app with new store
