+++ local/models/store/index.ts
@@ -1,7 +1,7 @@
 import { useAccessCodesStore } from '~/store/accessCodes'
 import { useAccountsStore } from '~/store/accounts'
 import type { useUsersStore } from '~/store/users'
-import { useCardsStore } from '~/store/piniaCards'
+import { useCardsStore } from '~/store/cards'
 
 export interface StoreType {
     users: useUsersStore
